Wrap TX in a table like abstraction, only for life of TX
This is a simple way to execute  HBase operations (get, put, delete...)
in an isolated transactional context. This avoids the client code to
create specifically the transactional table when simple operations are
done.

NOTE: This is just a proposal, and only the transactional get has been
implemented so far. The usage from the user point of view is described
here:

https://git.corp.yahoo.com/scalable-computing/omid/wiki/Transactional-User-Operations

Comments are welcome!

Change-Id: I2a590aac81e5b5bb0d4d855713f6c9b7ae1faf2b